Link Spotify Player with Chromecast device
Once authenticated with Spotify, fetch the list of available chromecast
devices. The Spotify player will now be initialized with an associated
device id. Once mounted, it will send a capture tab message to the
browser extension, which will then start streaming the audio to the
linked Chromecast device.

This commit also includes improved error handling with an error
boundary, as well as centralized handling of authentication errors,
which will redirect the user to the auth page.